Tasks_and_codes/requirements.txt
agno
groq
python-dotenv
pandas
numpy
ddgs
duckduckgo-search
yfinance
pypdf
lancedb
streamlit
tantivy 
openai 
sqlalchemy